Add files via upload
The Consumer Price Index (CPI) is a key indicator of production materials, derived from supply, and reflects the population's consumption of products. Therefore, demand also fluctuates over time. Both demand and supply change over time, and they determine prices in different geographic regions, which can vary. Additionally, the country’s exchange rate and currency unit are variable components that impact this. To standardize these factors, we consider the Consumer Price Index, which combines all of these elements and serves as a key indicator for assessing the significant impact on the cost of living in any location in the future, to maintain lifestyle standards.

Although earning potential and actual income are not always aligned, it is observed that despite having a higher or satisfactory income, an individual may not be able to save as much as they would like. On the other hand, people with lower incomes in areas with a lower CPI may save more. This judgment may apply to those with higher incomes, but it needs to be balanced with the cost of living in metropolitan areas. Therefore, the question is: Is a high income enough to beat the Consumer Price Index in a particular city or location? Or is the CPI low enough to allow someone to survive despite having a high income?